Develop

+++ src/task/task.service.ts
@@ -46,7 +46,11 @@ export class TaskService {
     }
   }
 
-  async getTasksById (id: string){
+  async getTasksById (id: string, userId: string){
+    const checkId = await this.taskModel.findOne({userId: userId})
+    if(!checkId){
+      throw new UnauthorizedException('Not permitted!')
+    }
     const find = await this.taskModel.findById({_id:id})
     if(!find){
       throw new NotFoundException("Task not found")
@@ -54,8 +58,12 @@ export class TaskService {
     return find;
   }
 
-  async updateTasks (id:string, updateTaskDto: UpdateTaskDto){
-    
+  async updateTasks (id:string, updateTaskDto: UpdateTaskDto, userId: string){
+
+    const find = await this.taskModel.findById({_id:id})
+    if(!(find.createdBy.toString()=== userId)){
+      throw new UnauthorizedException('You do not have access to update this task')
+    }
     const updateTask = await this.taskModel.findByIdAndUpdate(id, 
     {$set: updateTaskDto}, {new:true})
 
@@ -65,8 +73,12 @@ export class TaskService {
     return updateTask
   }
 
-  async deleteTasks (id:string){
+  async deleteTasks (id:string, userId:string){
 
+    const find = await this.taskModel.findById({_id:id})
+    if(!(find.createdBy.toString()=== userId)){
+      throw new UnauthorizedException('You do not have access to delete this task')
+    }
     const deleteTask = await this.taskModel.findByIdAndDelete({_id:id})
     if(!deleteTask){
       throw new NotFoundException("Task not found")
